Docstrings in cross_lines_z (#14)
* Clean up cross_lines_z scripts

* cross lines docstrings

* import fix

* more import fix
